[refactor] 디렉토리 구조에 맞게 작업 변경
