+++ client/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './common';
 export { default as LoginBox } from './LoginBox/LoginBox';
+export { default as ChannelListBox } from './ChannelListBox/ChannelListBox';
