+++ client/src/store/modules/index.ts
@@ -1,7 +1,9 @@
 import { combineReducers } from 'redux';
+import channelSlice, { CHANNELS } from './channels';
 import authSlice, { AUTH } from './auth';
 
 const rootReducer = combineReducers({
+  [CHANNELS]: channelSlice,
   [AUTH]: authSlice,
 });
